# Correction makes millis() exact for several speeds