--- a/README.md
+++ b/README.md
@@ -4,6 +4,7 @@
 This repo contains the Arduino corefiles used with [MightyCore](https://github.com/MCUdude/MightyCore), [MegaCore](https://github.com/MCUdude/MegaCore), [MiniCore](https://github.com/MCUdude/MiniCore) and [MajorCore](https://github.com/MCUdude/MightyCore).
 
 ## Supported devices
+
 * ATmega640, ATmega1280, ATmega2560
 * ATmega64, ATmega128, ATmega1281, ATmega2561
 * AT90CAN32, AT90CAN64, AT90CAN128
@@ -12,7 +13,9 @@ This repo contains the Arduino corefiles used with [MightyCore](https://github.c
 * ATmega8, ATmega48/P/PA/PB, ATmega88/P/PA/PB, ATmega168/P/PA/PB, ATmega328/P/PA/PB
 
 ## Supported clock frequencies
-By supported I mean clocks that accurate timing is implemented for (millis, micros, delay, delayMicroseconds).
+By supported I mean clocks that accurate timing is implemented for (millis,
+micros, delay, delayMicroseconds).
+
 * 32 MHz
 * 24 MHz
 * 20 MHz
@@ -28,3 +31,21 @@ By supported I mean clocks that accurate timing is implemented for (millis, micr
 * 2 MHz
 * 1.8432 MHz
 * 1 MHz
+
+### Accuracy of `millis()`
+
+For the clock speeds listed above, `millis()` is corrected to zero drift.
+Even for very long run times, the `millis()` function will precisely follow the
+oscillator used.
+We do not report the rollover of the `unsigned long` millis counter that occurs
+every day or two; such would have to be done in the user's program.
+Often this is not necessary:  Expressions like
+
+    (int) (millis() - millis_old)
+
+are correct even when rolling over provided `millis_old` is of type `unsigned long`
+and old and new time are no more than 16 seconds apart.
+
+For clock speeds of 16 MHz and below, the return value of `millis()`
+occasionally jumps up by more than one (notwithstanding zero long-time drift).
+Thus, when relying on consecutive returns, run at 18.432 MHz or above.