--- a/README.md
+++ b/README.md
@@ -13,6 +13,7 @@ This repo contains the Arduino corefiles used with [MightyCore](https://github.c
 * ATmega8, ATmega48/P/PA/PB, ATmega88/P/PA/PB, ATmega168/P/PA/PB, ATmega328/P/PA/PB
 
 ## Supported clock frequencies
+
 By supported I mean clocks that accurate timing is implemented for (millis,
 micros, delay, delayMicroseconds).
 
@@ -32,13 +33,13 @@ micros, delay, delayMicroseconds).
 * 1.8432 MHz
 * 1 MHz
 
-### Accuracy of `millis()`
+### Exactness of `millis()`
 
 For the clock speeds listed above, `millis()` is corrected to zero drift.
 Even for very long run times, the `millis()` function will precisely follow the
 oscillator used.
-We do not report the rollover of the `unsigned long` millis counter that occurs
-every day or two; such would have to be done in the user's program.
+We do not register the rollover of the `unsigned long` millis counter that
+occurs every day or two; such would have to be done in the user's program.
 Often this is not necessary:  Expressions like
 
     (int) (millis() - millis_old)
@@ -48,4 +49,4 @@ and old and new time are no more than 16 seconds apart.
 
 For clock speeds of 16 MHz and below, the return value of `millis()`
 occasionally jumps up by more than one (notwithstanding zero long-time drift).
-Thus, when relying on consecutive returns, run at 18.432 MHz or above.
+Thus, when relying on consecutive returns, run at 18 MHz or above.