PR originally created against MiniCore, now moved here.