@cburstedde 2020-12-15
> Haven't had the chance to look through this PR yet. Quick question though, will 32 MHz work? Even though none of my cores supports this frequency, the timing functions do.

millis () for 32 MHz is unaffected, since it was exact before already.