@cburstedde 2020-12-15
Done with 22.1184 (untested). I did not update the delay functions, just millis and micros.

Turns out 18.432 is actually the most recommended of the odd ones with micros() accuracy of 125ppm.