@cburstedde 2020-12-17
So I think I'm done with millis. There was a non-obvious roundoff in one of the typical F_CPU scaling macros, which accounted for several percent of error for some non-power-of-two clocks. Fixed this and added fine tuning with exactly zero long-term drift.

Tested with three different F_CPU settings on my laptop, with 18.432 MHz below 100ppm error by manual timing.