@cburstedde 2020-12-20
The millis() correction changes just the ISR. Its number of cycles will likely not be constant for most odd frequencies, but as you said that was never the case. The additional cost of this PR is an unsigned char compare and increment, which I'd judge acceptable going from an error in the percent range to zero. But please check it out and let me know what you think.

I've only got two crystals now, but I've tested several of the other odd frequencies by re#defining F_CPU. Seems alright.