@MCUdude 2020-12-21
Here are some results when testing with a signal generator at various frequencies. The way I know how long the ISR lasted is by setting PORTD = 0xFF as the first thing in the ISR, and setting PORTD = 0x00 as the last thing. Tested on an Atmega88, since this was a board I had within reach. But any AVR would do.

Here is how many instructions each ISR takes with including setting and clearing PORTD:

| F_CPU       | Cycles in ISR (old implementation) | Cycles in ISR (new implementation) |
|-------------|------------------------------------|------------------------------------|
| 1 MHz       | Jumping between 49 and 54          | Jumping between 51 and 52          |
| 1.8432 MHz  | Jumping between 49 and 54          | Jumping between 55, 59 and 61      |
| 2 MHz       | Jumping between 49 and 54          | Jumping between 51 and 52          |
| 3.6864 MHz  | ...                                | Jumping between 62, 63, 65 and 68  |
| 4 MHz       | ...                                | Jumping between 51 and 52          |
| 7.3728 MHz  |                                    | Jumping between 62, 65, 66 and 68  |
| 8 MHz       |                                    | Jumping between 51 and 52          |
| 11.0592 MHz |                                    | Jumping between 62, 63, 65 and 68  |
| 12 MHz      |                                    | Jumping between 55, 57, 59 and 61  |
| 14.7456 MHz |                                    | Jumping between 55 and 61          |
| 16 MHz      |                                    | Jumping between 51 and 52          |
| 18.4320 MHz |                                    | Jumping between 55 and 59          |
| 20 MHz      |                                    | Jumping between 55, 59, 63 and 66  |

Bottom line, this looks very consistent. Actually, the 20 MHz ISR was the one with the most jitter.

But how did you set up a way to long-term test the accuracy of millis()? Did you just print the millis() value to the serial monitor every minute or so, or did you do something else? One problem with this approach is that when the ISR fires while it's waiting (in a delay() or something) the microcontroller doesn't subtract the amount of time it took to execute the ISR from the delay function, and over time, timing will be slightly inaccurate.