@MCUdude 2020-12-21
> The 20 MHz formula is the weirdest, since it uses a counter that rolls on reaching 5 and an increment for odd counts (2 out of 5). I can ponder this one and see if I find something more elegant.

Great! I'll keep this open a little longer then. FYI I will be leaving my scope and my signal generator for a few days around Christmas. I'm able to continue testing tomorrow evening and a little bit on Wednesday morning.

> i did not use delay(), since I know that it misses cycles long-term. What I did is run a blink with a 3-second sequence based on millis() alone, and hit a time measurement command on my linux laptop like date "+%c %s.%N". With a bit of practice, the hit error is about 1/10 of a second. Then I take the difference between some current time and the initial time and divide that by 3. The fractional part of the result must be close to 0 +/- say 0.0xy, or something is wrong. It takes several hours for this error to grow closer to 1, when the lock-on gets lost. While still locked, the relative error is about 1. - ((tnow - tbegin) / 3.) / (integral part of ((tnow - tbegin) / 3.)), which approaches 100ppm after some dozen minutes.

Wow, that sounds like a very comprehensive testing method! Since your "setup" is a little difficult for me to set up (and wrap my head around, to be honest), are you confident that the other untested frequencies are reasonably accurate? I mean, you have already tested with 18.432 MHz, which is one of those "weird" baud rate friendly clocks, so you have already proven that your math works 👍