@cburstedde 2020-12-21
I've pushed an alternate 20 MHz to cburstedde@631a3495d1d4a6821ae5c7c4d93f83aa5659a99a.
How does this one behave?