@cburstedde 2020-12-21
The 20 MHz drift is well within spec. I'd say the benefit of removing the general roundoff beats the risk of overlooking one of the fine-tuned corrections. While only owning two crystals, I've checked many other frequencies by re#defining F_CPU and rescaling the measurements.