@MCUdude 2020-12-21
> I've pushed an alternate 20 MHz to cburstedde/MCUdude_corefiles@631a349.
How does this one behave?

Well, it now jumps between 58, 59, 62, 63, and 66 cycles to execute the ISR. So it now has five different execution times, but the maximum difference between the slowest and fastest is now 8 cycles instead of 11.